+++ packages/core/src/node/runtimeModule/types.ts
@@ -20,9 +20,9 @@ export enum RuntimeModuleID {
   GlobalComponents = 'virtual-global-components',
   Routes = 'virtual-routes',
   SiteData = 'virtual-site-data',
+  PageData = 'virtual-page-data',
   SearchIndexHash = 'virtual-search-index-hash',
   socialLinks = 'virtual-social-links',
   I18nText = 'virtual-i18n-text',
   SearchHooks = 'virtual-search-hooks',
-  RuntimeConfig = 'virtual-runtime-config',
 }
